test(git): split the relocate-path cases out
Rebased onto master, whose #2714 grew worktree_archive_test.go; these
additions pushed it to 1609 lines, over the 1500 limit. Split rather
than grandfathered (#1145): the fast-path-selection and bounded-step
cases move to worktree_relocate_bounds_test.go, and the round-trip and
cleanup-failure cases stay put. No test bodies changed.
